requirements.txt
amqp==2.6.1
asgiref==3.5.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==4.4.7
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.13
django-celery-beat==2.2.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==4.6.11
lxml==4.8.0
MarkupSafe==2.1.1
packaging==21.3
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
vine==1.3.0
gunicorn==20.0.4
redis==4.3.1
